+++ application/models/Pozorista_m.php
@@ -7,6 +7,15 @@ class Pozorista_m extends CI_Model {
     }
 
     public function insert($creatorUsername, $fileName) {
+        if (!$creatorUsername
+                || !$this->input->post('naziv')
+                || !$this->input->post('adresa')
+                || !$this->input->post('telefon')
+                || !$this->input->post('email')
+                || !$this->input->post('opis')
+        ) {
+            return FALSE;
+        }
         $data = array(
             'Naziv' => $this->input->post('naziv'),
             'Adresa' => $this->input->post('adresa'),
@@ -15,12 +24,14 @@ class Pozorista_m extends CI_Model {
             'Slika' => $fileName,
             'Opis' => $this->input->post('opis'),
             'Username' => $creatorUsername
-       
         );
         return $this->db->insert('pozoriste', $data);
     }
 
     public function findOne($id) {
+        if (!$id) {
+            return FALSE;
+        }
         $query = $this->db->get_where('pozoriste', array('PozID' => $id));
         return $query->row_array();
     }
@@ -30,24 +41,38 @@ class Pozorista_m extends CI_Model {
         $query = $this->db->get('pozoriste');
         return $query->result_array();
     }
-    
-    public function findNaziv($id){
+
+    public function findNaziv($id) {
+        if (!$id) {
+            return FALSE;
+        }
         $this->db->select('Naziv');
         $query = $this->db->get_where('pozoriste', array('PozID' => $id));
         return $query->row_array();
     }
 
     public function removeOne($id) {
+        if (!$id) {
+            return FALSE;
+        }
         $this->db->delete('pozoriste', array('PozId' => $id));
     }
 
     public function update($fileName) {
+        if (!$this->input->post('naziv')
+                || !$this->input->post('adresa')
+                || !$this->input->post('telefon')
+                || !$this->input->post('email')
+                || !$this->input->post('opis')
+        ) {
+            return FALSE;
+        }
         $data = array(
             'Naziv' => $this->input->post('naziv'),
             'Adresa' => $this->input->post('adresa'),
             'Telefon' => $this->input->post('telefon'),
-            'Email' => $this->input->post('email'),       
-            'Opis' => $this->input->post('opis'),       
+            'Email' => $this->input->post('email'),
+            'Opis' => $this->input->post('opis'),
         );
         if ($fileName) {
             $data['Slika'] = $fileName;
@@ -57,4 +82,3 @@ class Pozorista_m extends CI_Model {
     }
 
 }
-
